requirements.txt
anyio==3.6.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
fastapi==0.88.0
flake8==6.0.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
numpy==1.23.5
packaging==21.3
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.10
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
